Moved csv file generation to a new business layer
(cherry picked from commit 7cbdc5b092bf7ab0f86610ec5cfa0eb95f9c6c3b)
